FLAPPY BİRD CLONE GAME
This project contains a clone of the classic Flappy Bird game. The main objective of the game is to keep the bird flying for as long as possible by navigating it through gaps between pipes. The project is developed using the Java programming language and the LibGDX game development library.